Add BMI calculator functionality
